BatoTo (v4) | HOTFIX: Reorder server fallbacks (#12453)
BatoTo(v4) | HOTFIX: Reorder server fallbacks

The current implementation lists the k subdomains first and iterates
through the subdomains until it finds a successful image url. The
problem is that right now the k subdomains are mostly failing and it
takes 5 second to timeout, meaning it takes a long time to get a
successful image url and thus load times can be over a minute.

We do this quick fix to reorder the subdomains to have the n subdomains
first as those are mostly working. Ideally a more sophisticated solution
is done, but as this is a problem happening now, we do the simplest,
quickest fix to get this out as soon as possible.